+++ src/types.ts
@@ -5,5 +5,5 @@ export type Brightness = number
 export type DateFormatted = string
 export type HoursAndMinutes = string
 export type BatteryLevel = number
-export type Data = [DateFormatted, IsCharging, Brightness, IsLidCLosed]
-export type Log = [DateFormatted, DateFormatted, HoursAndMinutes]
+export type Data = [DateFormatted, IsCharging, Brightness, IsLidCLosed, BatteryLevel]
+export type Log = [DateFormatted, DateFormatted, HoursAndMinutes, BatteryLevel, BatteryLevel]